[+] move maths to shared, change osu parser name, fix minor issues
