+++ osupyparser/osu/__init__.py
@@ -1,3 +1,3 @@
 from __future__ import annotations
 
-from .parser import OsuFile
+from .parser import OsuBeatmapFile
