+++ osupyparser/shared/maths.py
